settings: Move function to set up folder widget dropdown.
